+++ mule/ci/__main__.py
@@ -1,6 +1,6 @@
 import sys
 
-from mule.poc.driver import main
+from mule.ci.driver import main
 
 if __name__ == "__main__":
     sys.exit(main())
