+++ mule/task/mule.py
@@ -1,12 +1,10 @@
 import os
-import sys
-
 
 import mule
-from mule.error import messages
+from mule.task.error import messages
 from mule.util import file_util, update_mule_file
 import mule.util.yaml.env_var_loader as yaml_util
-import mule.validator
+import mule.task.validator
 
 
 def _read_mule_yamls(mule_yamls):
